chore(): cleanup

+++ superawesome-base/src/main/java/tv/superawesome/lib/sametrics/models/SAPerformanceMetricModel.java
@@ -1,11 +1,11 @@
 package tv.superawesome.lib.sametrics.models;
 
 public class SAPerformanceMetricModel {
-  public final Double value;
+  public final Long value;
   public final SAPerformanceMetricName metricName;
   public final SAPerformanceMetricType metricType;
 
-  public SAPerformanceMetricModel(double value,
+  public SAPerformanceMetricModel(Long value,
                                   SAPerformanceMetricName metricName,
                                   SAPerformanceMetricType metricType) {
     this.value = value;
